created patch and delete mappings

+++ fictionfood/src/main/java/com/portfolio/fictionfood/recipe/RecipeController.java
@@ -1,5 +1,6 @@
 package com.portfolio.fictionfood.recipe;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.portfolio.fictionfood.exception.UnauthorizedException;
 import com.portfolio.fictionfood.user.User;
 import com.portfolio.fictionfood.user.UserRepository;
@@ -17,7 +18,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.Map;
 
@@ -27,6 +27,7 @@ import java.util.Map;
 public class RecipeController {
 
     private static final Logger logger = LoggerFactory.getLogger(RecipeController.class);
+    private final ObjectMapper objectMapper;
     private final RecipeService recipeService;
     private final RecipeRepository recipeRepository;
     private final UserRepository userRepository;
@@ -71,17 +72,18 @@ public class RecipeController {
     @PostMapping
     public ResponseEntity<?> postRecipe(@RequestPart("recipe") String recipeJson,
                                         @RequestPart("image") MultipartFile image,
-                                        Principal principal) {
-        String name = principal.getName();
-        logger.info("Attempting to post a new recipe by user: {}", name);
+                                        @AuthenticationPrincipal User currentUser) {
+
+        logger.info("Attempting to post a new recipe by user: {}", currentUser.getUsername());
 
         if (!checkIfAllowedToPost(SecurityContextHolder.getContext().getAuthentication())) {
-            logger.warn("User: {} is not allowed to post more recipes", name);
+            logger.warn("User: {} is not allowed to post more recipes", currentUser.getUsername());
             return new ResponseEntity<>(HttpStatus.TOO_MANY_REQUESTS);
         }
 
         try {
-            Recipe recipe = recipeService.postRecipe(recipeJson, image, principal);
+            PostRecipeDto recipeDto = objectMapper.readValue(recipeJson, PostRecipeDto.class);
+            Recipe recipe = recipeService.postRecipe(recipeDto, image, currentUser);
             logger.info("Recipe successfully posted with ID: {}", recipe.getId());
             return ResponseEntity.status(HttpStatus.CREATED).body("Recipe successfully created!");
         } catch (IOException e) {
@@ -90,9 +92,44 @@ public class RecipeController {
         }
     }
 
+    @PatchMapping
+    public ResponseEntity<?> updateRecipe(@RequestPart String recipeJson,
+                                          @RequestPart MultipartFile image,
+                                          @AuthenticationPrincipal User currentUser) {
+        try {
+            Recipe updatedRecipe = objectMapper.readValue(recipeJson, Recipe.class);
+            if (!recipeRepository.findById(updatedRecipe.getId()).orElseThrow().getAuthor().equals(currentUser)) {
+                logger.warn("User: {} tried to edit a recipe which did not belong to them", currentUser.getUsername());
+                throw new UnauthorizedException("You are not authorized to edit this recipe");
+            }
+            Recipe recipe = recipeService.updateRecipe(updatedRecipe, image);
+            logger.info("Recipe successfully edited with ID: {}", recipe.getId());
+            return ResponseEntity.status(HttpStatus.OK).body("Recipe successfully updated!");
+        } catch (UnauthorizedException u) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(u);
+        } catch (IOException b) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(b);
+        } catch (Error e) {
+            logger.error("Error editing recipe", e);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e);
+        }
+    }
+
     @DeleteMapping("/{id}")
-    public void deleteRecipeById(@PathVariable("id") long id) {
-        recipeRepository.deleteById(id);
+    public ResponseEntity<?> deleteRecipeById(@PathVariable("id") long id, @AuthenticationPrincipal User currentUser) {
+        try {
+            if (!recipeRepository.findById(id).orElseThrow().getAuthor().equals(currentUser) && !currentUser.getRole().equals(UserRole.MODERATOR)) {
+                recipeRepository.deleteById(id);
+                logger.info("Recipe successfully deleted with ID: {}", id);
+                throw new UnauthorizedException("You are not authorized to edit this recipe");
+            }
+            return ResponseEntity.status(HttpStatus.OK).body("Recipe successfully deleted!");
+        } catch (UnauthorizedException u) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(u);
+        } catch (Error e) {
+            logger.error("Error editing recipe", e);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e);
+        }
     }
 
     public boolean checkIfAllowedToPost(Authentication authentication) {
